Upgrade LanguageClient to 8.1.0
Upgrade LanguageClient to 8.1.0
- Update code to use `start` instead of `onReady`: this is not a trivial as it sounds as behaviors changed slightly, particularly in regards to error handling.
- Adding `initializationFailedHandler` to the `LanguageClient` created by `createClient` so preexisting connection errors (e.g. `Pending response rejected since connection got disposed`) do not surface as notifications (chane introduced in LC 3.17).
- Wrapping the `LanguageClient` created by `createClient` with an implementation that overrides `error` to override the `force` flag that causes some errors to bubble up.  There were and are ongoing reports about this: 8.0 allowed errors to bubble up which exposed the issues with `force`, and `8.1.0` fix some, but not all issues (and still there as of latest version on 2/28/2025).

Misc:
- Move `SorbetLanguageClient` to `sorbetLanguageClient.ts` to match class name.
- `languageClient.ts`, next to `languageClient.metrics.ts` operate on `LanguageClient`. Conviniently, existing `languageClient.test.ts` matches this separation.

References
- Previous attempt: Revert "update vscode-languageclient with LSP 3.17 support (#6492)" #6501
- Relevant spec changes:
   - `onReady` removed: [3.17.0](https://github.com/microsoft/vscode-languageserver-node/blob/main/README.md#3170-protocol-800-json-rpc-800-client-and-800-server)
   - Updates to error handling (required but not enough to limit spurious errors from  showing as notifications, unless handled) : [3.17.3](https://github.com/microsoft/vscode-languageserver-node/blob/main/README.md#3173-protocol-810-json-rpc-810-client-and-810-server)
